requirements.txt
discord.py==1.7.3
toml
asyncpg
git+https://github.com/Rapptz/discord-ext-menus
# Wand # ????
